drivers/ao_packet: Switch time values to AO_TICK_TYPE
Make sure times are kept in the right type. Also resolved other
-Wconversion issues.

+++ src/drivers/ao_packet_slave.c
@@ -21,7 +21,7 @@
 static void
 ao_packet_slave(void)
 {
-	ao_tx_packet.addr = ao_serial_number;
+	ao_tx_packet.addr = (uint8_t) ao_serial_number;
 	ao_tx_packet.len = AO_PACKET_SYN;
 	ao_packet_restart = 1;
 	while (ao_packet_enable) {
